requirements.txt
# Core AI
google-generativeai==0.3.2
numpy==1.24.3

# Google Drive API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Document Processing
python-docx==0.8.11
python-pptx==0.6.21
pymupdf==1.23.8
openpyxl==3.1.2

# OCR Support
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78

# Google Vision API (optional, pour OCR plus précis)
google-cloud-vision==3.4.4

# Vector Storage (Simple)
sqlite3  # Built-in Python
pickle   # Built-in Python

# Template Generation
docxtpl==0.16.7
jinja2==3.1.2

# Utils
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.65.0
pandas==2.0.3

# Optional: Web interface simple
streamlit==1.28.1